Adding dictionaries feature & cleaning up inventory (#8)
* Changed defaults and cleaned up the comments

The inventory file is now so that you can use it without the need of the README file, but it also doesn't contain unnecessary lines. Those I've move to the readme file.

* Sync with inventory

I made sure that the inventory file contains all the necessary info for setup. The README is now just a walkthrough with some extra information.

* Update README.md

* right url markup

* added option to choose dictionaries

* Update inventory

* restic_repo

* restic is available